When reviewing a version, show data directly from the version
