+++ app/decorators/seller_version_decorator.rb
@@ -1,0 +1,13 @@
+class SellerVersionDecorator < BaseDecorator
+
+  def agreed_by_email
+    agreed_by.email if agreed_by.present?
+  end
+
+  def addresses
+    super.map {|address|
+      SellerAddressDecorator.new(address, view_context)
+    }
+  end
+
+end
