+++ app/views/ops/seller_versions/_layout.html.erb
@@ -4,7 +4,7 @@
       <aside>
         <h1>
           <span class="type">Seller application</span>
-          <%= application.seller.name %>
+          <%= application.name %>
         </h1>
         <nav class="vertical">
           <ul>
